+++ familyhelper-assets-sdk/src/main/java/com/dwarfeng/familyhelper/assets/sdk/util/Constants.java
@@ -17,6 +17,8 @@ public final class Constants {
     public static final int PERMISSION_LEVEL_GUEST = 1;
     public static final int PERMISSION_LEVEL_MODIFIER = 2;
 
+    public static final int IO_TRANS_BUFFER_SIZE = 4096;
+
     private Constants() {
         throw new IllegalStateException("禁止实例化");
     }