+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 #### 功能构建
 
+- 增加服务方法。
+  - com.dwarfeng.familyhelper.assets.stack.service.ItemFileOperateService。
+
 - 日志功能优化。
   - 优化默认日志配置，默认配置仅向控制台输出 `INFO` 级别的日志。
   - 优化日志配置结构，提供 `conf/logging/settings.xml` 配置文件及其不同平台的参考配置文件，以供用户自定义日志配置。